Start splitting up app.py
